feat(api): add SessionNotFinalizedError in unfinalize-session.js

+++ api/src/certification/session-management/domain/usecases/unfinalize-session.js
@@ -5,14 +5,15 @@
 
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
 import { NotFoundError } from '../../../../shared/domain/errors.js';
-import { SessionAlreadyPublishedError } from '../errors.js';
+import { SessionAlreadyPublishedError, SessionNotFinalizedError } from '../errors.js';
 
 /**
  * @param {object} params
  * @param {SessionManagementRepository} params.sessionManagementRepository
  * @param {FinalizedSessionRepository} params.finalizedSessionRepository
+ * @throws {NotFoundError} the session does not exist or its access is restricted
  * @throws {SessionAlreadyPublishedError} the session is already published
- * @throws {NotFoundError} the finalized session does not exist or its access is restricted
+ * @throws {SessionNotFinalizedError} the session is not finalized
  */
 const unfinalizeSession = async function ({ sessionId, sessionManagementRepository, finalizedSessionRepository }) {
   return DomainTransaction.execute(async () => {
@@ -24,6 +25,10 @@ const unfinalizeSession = async function ({ sessionId, sessionManagementReposito
 
     if (session.isPublished()) {
       throw new SessionAlreadyPublishedError();
+    }
+
+    if (!session.isFinalized()) {
+      throw new SessionNotFinalizedError();
     }
 
     await finalizedSessionRepository.remove({ sessionId });
